binarize: clean up formatting and add invert flag.
The existing qr_binarize routine would always invert images (i.e. light
pixels would become black and dark pixels would become white), and the
decoding algorithm depends on this inversion. However, it is useful
for this to be configurable for use as a general purpose binarization
routine and to give the user more flexibility of input (e.g. maybe the
source images are already inverted for some reason).