Move example/test code into example directory
reader.c and image.c are standalone source files that compose a test
executable (that also conveniently serve as an example for basic API
usage). This change also caught a spurious (but harmless) include in
binarize.c, which has been moved.

Also the test code build in the makefile was kind of crusty and it is
now somewhat less crusty. While I was at it, I also told clang to
quiet down about the operator precedence warnings. I spent a little
bit of time trying to fix these a while ago but pretty quickly aborted
that attempt, as it was way too easy to break the decoder. I think
even clang-tidy would break it, though it has been a number of years
since I last tried.